Fix monster spawn packet layout and make the LogFile option work
SWPACKET_IN_INFO_MONSTER1 was missing a 4-byte field ahead of
data1Count, and SWPACKET_IN_INFO_MONSTER2 a trailing float, so
data1Count was read from +78 instead of +82. Any spawn packet
carrying more than one monster desynced the offset walk and dropped
every entry after the first; single-monster packets hid it because
the misread byte happened to be 0. Verified against captured 0415
and 0423 frames - the corrected layout consumes each to the byte.

Widen the offset accumulator from short to uint32_t so a large batch
cannot wrap it negative.

Log::Enable() had no callers, so LogFile=true set shouldLog but never
opened the file and nothing was ever written.

+++ Soulworker Utility/Soulworker Packet/SWPacketOtherInfosMonster.cpp
@@ -12,7 +12,7 @@ SWPacketOtherInfosMonster::SWPacketOtherInfosMonster(SWHEADER* swheader, BYTE* d
 void SWPacketOtherInfosMonster::Do() {
 
 	short counts = *(_data + sizeof(SWHEADER));
-	short offset = sizeof(SWHEADER) + sizeof(counts);
+	uint32_t offset = sizeof(SWHEADER) + sizeof(counts);
 	for (short i = 0; i < counts; i++) {
 
 		SWPACKET_IN_INFO_MONSTER1* pktHeader = (SWPACKET_IN_INFO_MONSTER1*)(_data + offset);
